Migrate the skills loop onto the typed path, behind a flag
Deliverables 2, 3, 4, 6 and 7 in one commit. They were meant to be separate,
and are not, because they are mutually referential: the migrated loop consults
the review surface for the re-proposal cooldown, and the review surface reads
the proposals the loop induces. Splitting them would have produced
intermediate commits that do not compile, which is worse than one honest
message. What each part does, and why:

--- 2. TYPED OBSERVATION EXTRACTION, AND THE REDACTOR IT NEEDS ---

Spec §5.6 says secrets are redacted before persistence, indexing, or
embedding. Nothing on the reflection path did that, and there was no
general-purpose text redactor in the workspace at all. The closest,
stella-observatory's settings scrubber, handles STRUCTURED TOML/JSON by key
name and lives in a read-only dashboard crate — it cannot see a credential
pasted into a sentence, which is the only shape a reflection lesson has.

context.db's own schema documents the consequence: it "holds recalled
memories, episodes (verbatim copies of past user prompts), and facts mined
from workspace content, so whatever secrets the workspace contains can end up
inside it", and narrows the file to 0600 in response. Permissions mitigate a
leak; they do not stop the secret being stored, and they do nothing for
reflections.jsonl, which is plain JSONL. This is a defect being closed.

No regex — not a workspace dependency, and a poor trade for a scanner this
small; the same reasoning already produced a hand-written FNV in mining.rs.
Half the tests assert that real lessons, git SHAs, paths and issue URLs pass
through UNTOUCHED: a redactor that eats real content gets turned off, which
leaks everything.

Replay-idempotence has two mechanisms and only one is the correctness claim. A
cursor bounds the work. Content-derived record ids make it *correct*, because
the cursor cannot be advanced atomically with the record write — a crash
between them would duplicate the batch or lose it. Tested by discarding the
cursor and re-running.

Task identity, and its limit, stated rather than buried: reflection runs once
per turn and stamps every lesson with one occurred_at, so `turn:<ts>` is a
real turn identity and blocks the concrete poisoning attack. It is NOT a task
boundary, and that is the unsafe direction — three turns on one task read as
three tasks. ReflectionLesson gains an optional task_id so a caller that knows
better can supply one without a log-format change; nothing populates it yet.

--- 3. PROPOSAL INDUCTION OVER THE UNCHANGED MINER ---

mine_skill_candidates is called verbatim, with unchanged thresholds, over
observations carrying the same text/reference/occurred_at/domains the lexical
path fed it. Not laziness: the candidate id IS the skill filename, so
re-deriving it differently orphans every file a user has, and the rendered
`## Evidence` lines quote reference and occurred_at directly, so changing
either moves bytes in files already on disk. occurred_at is recovered by
parsing source_ref rather than reformatting observed_at, for that reason.

What is added is the durable proposal: supporting observations, the
distinct-task count, and the scored components — reviewable before a skill
file exists. Re-inducing the same candidate with new evidence appends a
REVISION in the same lineage, so a decline can be remembered against it.

--- 4. THE MIGRATION ITSELF ---

Behind context.lifecycle.enabled, default off. The lexical path is kept
reachable rather than deleted, and that is a deliberate testing decision, not
timidity: a migration that deletes the thing it must stay compatible with has
no baseline left to compare against. Every spec §8 guarantee now runs under
BOTH loops via each_path(), and a divergence shows up as one path passing and
the other failing. None of the pre-existing assertions were edited to
accommodate the typed path — that edit would have been the regression.

Added: both loops must render BYTE-IDENTICAL skill files from the same
evidence, compared as bytes rather than field-wise, because a field-wise
comparison passes while a stray newline shows up in every user's git diff.

The cap, the no-clobber guard and the tombstone sweep are literally shared
code between the paths now (write_candidates, forgotten_texts), so "the same
filter" is a true statement rather than an aspiration.

--- 6. SOLO ADVISORY GOVERNANCE ---

Enforced by construction, not by call-site discipline. Every review decision
is Advisory or nothing; PromotionEventRecord::new refuses a system-granted
blocking enforcement outright, so "no inferred directive reaches blocking by
ANY path" is checkable — there is one path. Nothing writes a sharing scope at
all. The re-proposal cooldown reads the promotion event log rather than a
status column, so declining is reversible: a later Keep wins and both acts
stay readable.

--- 7. REVIEW SURFACE ---

`stella proposals list | keep | edit | ignore | refresh`. Every outcome is an
immutable event; state is a fold over the log, never a stored column, which is
what makes "replay identically" meaningful. `refresh` exists because "show me
what is pending" is useless if the only way to populate it is to run a turn
and hope.

One real bug found and fixed while testing this: the replay fold used
records_of_kind, which orders by (observed_at, record_id). Two decisions in
the same second carry the same observed_at, so the tiebreak was a content
hash — "ignore then keep" folded to `ignored` about half the time, and
DETERMINISTICALLY so, which is worse than flaky. Promotion events now read in
ledger append order (rowid), which is the actual authority for what happened
last in an append-only log.

Verified end-to-end against a real workspace, not only by test: a GitHub token
in a lesson reaches context.db as "[redacted]" and the raw token appears
nowhere in the database file; the listing prints "3 distinct task(s)"; ignore
hides it and survives a refresh; keep reverses it; and sqlite refuses a DELETE
against context_records.

Refs #714

+++ stella-core/src/context_record/lifecycle.rs
@@ -409,7 +409,7 @@ impl PromotionEventRecord {
 pub fn confidence_from_score(score: &ProposalScore) -> Result<Confidence, RecordValidationError> {
     // 20 per distinct task (the axis that resists poisoning), 5 per occurrence
     // beyond the first, 15 for an independently salient signal.
-    let tasks = u32::from(score.distinct_tasks.min(4)) * 20;
+    let tasks = score.distinct_tasks.min(4) * 20;
     let extra = score.occurrences.saturating_sub(1).min(4) * 5;
     let salient = if score.salient { 15 } else { 0 };
     Confidence::new((tasks + extra + salient).min(100) as u16)
